+++ sqlstore/src/main/resources/db/migration/V8501__requisitionId.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Sample ADD COLUMN requisitionId varchar(50);